+++ doorstop/core/tree.py
@@ -9,9 +9,9 @@ import subprocess
 import logging
 
 from doorstop.core.base import BaseValidatable
-from doorstop.core.types import ID
+from doorstop.core.types import Prefix, ID
 from doorstop.common import get_tree, DoorstopError, DoorstopWarning
-from doorstop.core.document import Document, get_prefix
+from doorstop.core.document import Document
 from doorstop.core import vcs
 
 
@@ -197,11 +197,11 @@ class Tree(BaseValidatable):  # pylint: disable=R0902
 
     @clear_document_cache
     @clear_item_cache
-    def new_document(self, path, prefix, sep=None, digits=None, parent=None):  # pylint: disable=R0913
+    def new_document(self, path, value, sep=None, digits=None, parent=None):  # pylint: disable=R0913
         """Create a new document and add it to the tree.
 
         @param path: directory path for the new document
-        @param prefix: document's prefix (or document)
+        @param value: document or prefix
         @param sep: separator between prefix and numbers
         @param digits: number of digits for the document's numbers
         @param parent: parent document's prefix
@@ -211,7 +211,7 @@ class Tree(BaseValidatable):  # pylint: disable=R0902
         @return: newly created and placed Document
 
         """
-        prefix = get_prefix(prefix)
+        prefix = Prefix(value)
         document = Document.new(self,
                                 path, self.root, prefix, sep=sep,
                                 digits=digits, parent=parent)
@@ -228,10 +228,10 @@ class Tree(BaseValidatable):  # pylint: disable=R0902
         return document
 
     @clear_item_cache
-    def add_item(self, prefix, level=None):
+    def add_item(self, value, level=None):
         """Add a new item to an existing document by prefix.
 
-        @param prefix: document's prefix (or document)
+        @param value: document or prefix
         @param level: desired item level
 
         @raise DoorstopError: if the item cannot be created
@@ -239,7 +239,7 @@ class Tree(BaseValidatable):  # pylint: disable=R0902
         @return: newly created Item
 
         """
-        prefix = get_prefix(prefix)
+        prefix = Prefix(value)
         document = self.find_document(prefix)
         self.vcs.lock(document.config)  # prevents duplicate item IDs
         item = document.add_item(level=level)
@@ -333,17 +333,17 @@ class Tree(BaseValidatable):  # pylint: disable=R0902
         # Return the item
         return item
 
-    def find_document(self, prefix):
+    def find_document(self, value):
         """Get a document by its prefix.
 
-        @param prefix: document's prefix (or document)
+        @param value: document or prefix
 
         @raise DoorstopError: if the document cannot be found
 
         @return: matching Document
 
         """
-        prefix = get_prefix(prefix)
+        prefix = Prefix(value)
         logging.debug("looking for document '{}'...".format(prefix))
         try:
             document = self._document_cache[prefix]